move api getListProduct to HomeController
