+++ Website_Ecommerce.API/bin/Debug/net6.0/Website_Ecommerce.API.staticwebassets.runtime.json
@@ -1,1 +1,1 @@
-{"ContentRoots":["/Users/vovanban/Documents/Kỳ 7/PBL6_ECOMMERCE/Website_Ecommerce.API/wwwroot/"],"Root":{"Children":{"avatar.jpg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"avatar.jpg"},"Patterns":null},"logo.png":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"logo.png"},"Patterns":null}},"Asset":null,"Patterns":[{"ContentRootIndex":0,"Pattern":"**","Depth":0}]}}
+{"ContentRoots":["/Users/vovanban/Documents/Kỳ 7/PBL6_ECOMMERCE/Website_Ecommerce.API/wwwroot/"],"Root":{"Children":{"avatar.jpg":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"avatar.jpg"},"Patterns":null},"logo.png":{"Children":null,"Asset":{"ContentRootIndex":0,"SubPath":"logo.png"},"Patterns":null}},"Asset":null,"Patterns":[{"ContentRootIndex":0,"Pattern":"**","Depth":0}]}}
